feat: migrate the slack webhooks methods into the modules.slack package
